DemonioRegistroIncidentes/requirements.txt
pika>=1.3.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0